Non-eq calculation is available for CO only
Hello,

A really minor pull request. 

I notice it is possible to launch non-equilibrium calculation for any HITRAN molecule. I'm scared a random user wanting to do this for any other molecule than CO will be frustrated (and may argue that "it doesn't work"). Is it a good idea to append "(CO only)" to "Use non-equilibrium calculation" ?

Nicolas